web/pyproject.toml
[project]
name = "trackmaniabingo-web"
version = "0.1.0"
description = "Trackmania Bingo web frontend"
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
